// src/typings/User.ts
import { Group } from "./Group.ts";

// Generated by https://quicktype.io

export interface User {
    description: string;
    created: Date;
    isBanned: boolean;
    id: number;
    name: string;
    avatars: UserAvatars
}

// Generated by https://quicktype.io

export interface UserAvatars {
    bodyUrl: string;
    headshotUrl: string;
}

export interface UsernameSearch {
    username: string;
    id: number;
}


// Generated by https://quicktype.io

export interface UserGroups {
    group: Group;
    role: Role;
    isPrimaryGroup: boolean;
}


export interface Role {
    id: number;
    name: string;
    description: string;
    rank: number;
    memberCount: number;
}

// Generated by https://quicktype.io

export interface UserBadges {
    previousPageCursor: null;
    nextPageCursor: string;
    data: BadgeDatum[];
}

export interface BadgeDatum {
    id: number;
    name: string;
    description: string;
    enabled: boolean;
    iconImageId: number;
    awarder: {
        id: number;
        type: string;
    };
    statistics: {
        pastDayAwardedCount: number;
        awardedCount: number;
        winRatePercentage: number;
    };
    created: Date;
    updated: Date;
}


// Generated by https://quicktype.io

export interface UserSearch {
    previousPageCursor: null;
    nextPageCursor: null;
    data: SearchDatum[];
}

export interface SearchDatum {
    previousUsernames: string[];
    id: number;
    name: string;
    displayName: string;
}